requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bareunpy
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
fastapi==0.112.1
filelock==3.15.4
fsspec==2024.6.1
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.0
openai==0.28
packaging==24.1
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rsa==4.9
safetensors==0.4.4
setuptools==73.0.1
sniffio==1.3.1
starlette==0.38.2
sympy==1.13.2
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.23.0
websockets==13.0
